moving nested classes to an upper level eliminating double validation moving generated classes into a source directory
