+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/query/ParameterDescription.java
@@ -2,6 +2,7 @@ package com.gigaspaces.sql.aggregatornode.netty.query;
 
 import com.gigaspaces.sql.aggregatornode.netty.exception.ProtocolException;
 import com.gigaspaces.sql.aggregatornode.netty.utils.PgType;
+import com.gigaspaces.sql.aggregatornode.netty.utils.TypeUtils;
 import io.netty.buffer.ByteBuf;
 
 public class ParameterDescription extends TypeAware {
@@ -10,6 +11,6 @@ public class ParameterDescription extends TypeAware {
     }
 
     public <T> T read(Session session, ByteBuf buf, int format) throws ProtocolException {
-        return PgType.readParameter(session, buf, this, format);
+        return TypeUtils.readParameter(session, buf, this, format);
     }
 }
