+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/query/QueryProvider.java
@@ -10,11 +10,6 @@ import java.util.List;
  */
 public interface QueryProvider {
     /**
-     * Initializes query provider
-     */
-    void init() throws ProtocolException;
-
-    /**
      * Executes multiline query.
      *
      * @param session Session
