+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/query/Statement.java
@@ -1,6 +1,7 @@
 package com.gigaspaces.sql.aggregatornode.netty.query;
 
 import org.apache.calcite.sql.SqlNode;
+import org.apache.calcite.sql.validate.SqlValidator;
 
 public interface Statement extends AutoCloseable {
     /**
@@ -12,6 +13,8 @@ public interface Statement extends AutoCloseable {
      * @return Parsed query
      */
     SqlNode getQuery();
+
+    SqlValidator getValidator();
 
     /**
      * @return Parameters and result row descriptions
