+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/query/StatementDescription.java
@@ -1,6 +1,9 @@
 package com.gigaspaces.sql.aggregatornode.netty.query;
 
 public class StatementDescription {
+    public static final StatementDescription EMPTY =
+            new StatementDescription(ParametersDescription.EMPTY, RowDescription.EMPTY);
+
     private final ParametersDescription parametersDescription;
     private final RowDescription rowDescription;
