+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/server/MessageProcessor.java
@@ -414,8 +414,6 @@ public class MessageProcessor extends ChannelInboundHandlerAdapter {
     }
 
     private void onAuthenticationOK(ChannelHandlerContext ctx) throws ProtocolException {
-        queryProvider.init();
-
         ByteBuf buf = ctx.alloc().ioBuffer();
         writeAuthenticationOK(buf);
         writeParameterStatus(buf, "client_encoding", session.getCharset().name());
