+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/utils/TypeAny.java
@@ -1,0 +1,9 @@
+package com.gigaspaces.sql.aggregatornode.netty.utils;
+
+public class TypeAny extends PgType {
+    public static final PgType INSTANCE = new TypeAny();
+
+    public TypeAny() {
+        super(2276, "any", 4, 0, 0);
+    }
+}
