+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/utils/TypeCursor.java
@@ -1,0 +1,10 @@
+package com.gigaspaces.sql.aggregatornode.netty.utils;
+
+// TODO implement type encoder/decoder
+public class TypeCursor extends PgType {
+    public static final PgType INSTANCE = new TypeCursor();
+
+    public TypeCursor() {
+        super(1790, "refcursor", -1, 2201, 0);
+    }
+}
