+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/utils/TypeFloat8.java
@@ -1,0 +1,36 @@
+package com.gigaspaces.sql.aggregatornode.netty.utils;
+
+import com.gigaspaces.sql.aggregatornode.netty.exception.ProtocolException;
+import com.gigaspaces.sql.aggregatornode.netty.query.Session;
+import io.netty.buffer.ByteBuf;
+
+public class TypeFloat8 extends PgType {
+    public static final PgType INSTANCE = new TypeFloat8();
+
+    public TypeFloat8() {
+        super(701, "float8", 8, 1022, 0);
+    }
+
+    @Override
+    protected void asTextInternal(Session session, ByteBuf dst, Object value) throws ProtocolException {
+        TypeUtils.checkType(value, Double.class);
+        TypeUtils.writeText(session, dst, value.toString());
+    }
+
+    @Override
+    protected <T> T fromTextInternal(Session session, ByteBuf src) {
+        return (T) Double.valueOf(TypeUtils.readText(session, src));
+    }
+
+    @Override
+    protected void asBinaryInternal(Session session, ByteBuf dst, Object value) throws ProtocolException {
+        TypeUtils.checkType(value, Double.class);
+        dst.writeInt(8).writeDouble((Double) value);
+    }
+
+    @Override
+    protected <T> T fromBinaryInternal(Session session, ByteBuf src) throws ProtocolException {
+        TypeUtils.checkLen(src, 8);
+        return (T) Double.valueOf(src.readDouble());
+    }
+}
