+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/utils/TypeTimeTZ.java
@@ -1,0 +1,10 @@
+package com.gigaspaces.sql.aggregatornode.netty.utils;
+
+// TODO implement type encoder/decoder
+public class TypeTimeTZ extends PgType {
+    public static final PgType INSTANCE = new TypeTimeTZ();
+
+    public TypeTimeTZ() {
+        super(1266, "timetz", 12, 1270, 0);
+    }
+}
