+++ xap-extensions/xap-aggregator-node/src/main/java/com/gigaspaces/sql/aggregatornode/netty/utils/TypeUnknown.java
@@ -1,0 +1,9 @@
+package com.gigaspaces.sql.aggregatornode.netty.utils;
+
+public class TypeUnknown extends PgType {
+    public static final PgType INSTANCE = new TypeUnknown();
+
+    public TypeUnknown() {
+        super(705, "unknown", -2, 0, 0);
+    }
+}
