+++ xap-extensions/xap-jdbc/.gitignore
@@ -1,0 +1,2 @@
+# generated classes
+src/main/java/com/gigaspaces/jdbc/calcite/parser/generated
